Use float instead of uint since there are transactions that transfer more satoshis than 2^53. These floats are also bigger than Number.MAX_SAFE_INTEGER, but we have decimals anyway, so it should be fine

+++ src/Indexer.js
@@ -10,7 +10,7 @@ const TX_BATCH_SIZE = 20000;
 const SATOSHI = 100000000;
 
 const utxoValueSchema = new JSBinType({
-  'sats': 'uint',
+  'sats': 'float',
   'createdOnBlock': 'uint',
   'spentOnBlock?': 'uint',
   'spentInTx?': 'uint',
